Completing initial refactoring into modules
