+++ lib/catalog.js
@@ -4,13 +4,12 @@
 
 
 var catalog = function(){};
-var cleanup = null;
+var cleanup = require('./cleanup.js');
 var b64 = require('js-base64').Base64;
 var logger = require('./logger.js');
 var arrayUtil = require('./arrayUtil.js');
-catalog.prototype.updateCatalog = function(job, callback)
+catalog.prototype.updateCatalog = function(job)
 {
-        this.cleanup = callback;
         job.github.repos.getContent({
                 owner: job.config.owner,
                 repo: job.config.targetRepo,
@@ -21,16 +20,16 @@ catalog.prototype.updateCatalog = function(job, callback)
             var cat = b64.decode(err.content);
             job.catalog = JSON.parse(cat);
             job.catalog.sha = err.sha;
-            this.updateJSONCatalog(job);
+            updateJSONCatalog(job);
         }).catch(function (err) {
             logger.log("error downloading keynote-catalog.json",job,"Updating catalog", err);
             job.catalog = [];
             job.catalog.sha = "";
-            this.updateJSONCatalog(job);
+            updateJSONCatalog(job);
         })
 };
 
-    catalog.prototype.updateJSONCatalog = function(job) {
+function updateJSONCatalog(job) {
         var catEntry = null;
         if(job.catalog.sha === "")
         {
@@ -75,7 +74,7 @@ catalog.prototype.updateCatalog = function(job, callback)
             job.github.repos.createFile({owner:job.config.owner,repo:job.config.targetRepo,path:"keynote-catalog.json",message:job.config.commitMsg,content:catContent})
                 .then(function(err, res)
                 {
-                    this.updateMarkdownCatalog(job);
+                    updateMarkdownCatalog(job);
                 })
         }
         else
@@ -83,13 +82,13 @@ catalog.prototype.updateCatalog = function(job, callback)
             job.github.repos.updateFile({owner:job.config.owner, repo:job.config.targetRepo, path:"keynote-catalog.json",message:job.config.commitMsg,content:catContent,sha:job.catalog.sha})
                 .then(function(err,res)
                 {
-                    this.updateMarkdownCatalog(job);
+                    updateMarkdownCatalog(job);
                 })
         }
 
     }
 
-    catalog.prototype.updateMarkdownCatalog = function(job)
+function updateMarkdownCatalog(job)
     {
         var output = "";
         job.catalog.forEach(function(cat)
@@ -117,10 +116,10 @@ catalog.prototype.updateCatalog = function(job, callback)
             });
         });
         job.markdownCatalog = output;
-        this.uploadCatalog(job)
+        uploadCatalog(job)
     }
 
-    catalog.prototype.uploadCatalog = function (job)
+function uploadCatalog(job)
     {
         var b64content = b64.encode(job.markdownCatalog)
         job.github.repos.getContent({owner:job.config.owner,
@@ -135,8 +134,7 @@ catalog.prototype.updateCatalog = function(job, callback)
                     logger.log("Markdown catalog updated", job,"Finishing");
                     delete job.catalog;
                     delete job.markdownCatalog;
-                    //cleanup(job);
-                    this.cleanup.cleanup(job);
+                    cleanup.cleanup(job);
                 })
         }).catch(function(err){
             job.github.repos.createFile({owner:job.config.owner,repo:job.config.targetRepo,path:"keynote-catalog.md",message:job.config.commitMsg,content:b64content})
@@ -146,7 +144,7 @@ catalog.prototype.updateCatalog = function(job, callback)
                     delete job.catalog;
                     delete job.markdownCatalog;
                     //cleanup(job);
-                    this.cleanup.cleanup(job);
+                    cleanup.cleanup(job);
                 })
 
         })
