+++ lib/cleanup.js
@@ -6,6 +6,7 @@ var format = require('date-fns/format');  //https://github.com/date-fns/date-fns
 var differenceInMilliseconds = require('date-fns/difference_in_milliseconds'); //https://github.com/date-fns/date-fns
 var http = require('http');
 var logger = require('./logger.js');
+var exec = require('child_process').exec;
 
 var cleanup = function() {};
